+++ stac/CHANGELOG.md
@@ -6,9 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- `geo` feature ([#178](https://github.com/gadomski/stac-rs/pull/178))
+
 ### Changed
 
 - `Links::remove_relative_links` has the same vibe as `Links::remove_structural_links` ([#176](https://github.com/gadomski/stac-rs/pull/176))
+- Use our own `Geometry` structure ([#178](https://github.com/gadomski/stac-rs/pull/178))
 
 ## [0.5.0] - 2023-06-27